tests/wpt/web-platform-tests/storage/persist-permission-manual.https.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>StorageManager: permission state is granted</title>
    <p>Clear all persistent storage permissions before running this test.</p>
    <p>Test passes if there is a permission prompt and click allow store persistent data</p>
    <meta name="help" href="https://storage.spec.whatwg.org/#dom-storagemanager-persist">
    <meta name="author" title="Mozilla" href="https://www.mozilla.org">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    promise_test(function(t) {
      return navigator.storage.persist()
      .then(function(result) {
        assert_true(result);
        return navigator.storage.persisted();
      })
      .then(function(result) {
        assert_true(result);
      })
    }, 'Expect permission state is granted after calling persist()');
    </script>
  </body>
</html>
